further seperated the classes into DialectTraverser and DialectGenerate. Added tests for processFunction
